+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 v0.1.4 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix DLQMetadata decoding to use DLQMetadataDecoder func instead of inferred decoding from TopicType.
 
 
 v0.1.3 (2018-03-09)